groupe-protagoras/requirements.txt
langchain-core>=0.1.0
langchain-community>=0.1.0
pydantic>=1.10.0
jpype1>=1.6.0